+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -15,6 +15,7 @@ export default function apmApiIntegrationTests({
     loadTestFile(require.resolve('./alerts'));
     loadTestFile(require.resolve('./custom_dashboards'));
     loadTestFile(require.resolve('./dependencies'));
+    loadTestFile(require.resolve('./environment'));
     loadTestFile(require.resolve('./error_rate'));
     loadTestFile(require.resolve('./data_view'));
     loadTestFile(require.resolve('./correlations'));
